requirements.txt
python-dotenv==1.0.1
setuptools==75.6.0
twine==6.0.1

fastapi==0.115.8
starlette==0.45.3
sqlmodel==0.0.22
orjson==3.10.15
uvicorn==0.34.0